Refactor: Simplify base classes (#370)
* Refactor: Drop BaseInjector

* Drop base lambda handler

* Drop BaseHandleRequestParams

No longer used.

* Rename base.ts to ErrorCode.ts since it's the only file left.

* Silo handler functions (#371)

Move entrypoints for each lambda to individual files
