+++ bin/stacks/step-function-stack.ts
@@ -39,7 +39,7 @@ export class StepFunctionStack extends cdk.NestedStack {
     const checkStatusFunction = new NodejsFunction(this, `${SERVICE_NAME}-${stage}-CheckOrderStatusLambda`, {
       runtime: aws_lambda.Runtime.NODEJS_18_X,
       role: props.lambdaRole,
-      entry: path.join(__dirname, '../../lib/handlers/index.ts'),
+      entry: path.join(__dirname, '../../lib/handlers/check-order-status/index.ts'),
       handler: 'checkOrderStatusHandler',
       memorySize: 512,
       bundling: {
