+++ lib/handlers/base/api-handler.ts
@@ -8,7 +8,8 @@ import {
 
 import { default as bunyan, default as Logger } from 'bunyan'
 import Joi from 'joi'
-import { BaseHandleRequestParams, BaseInjector, BaseLambdaHandler, BaseRInj, ErrorCode } from './base'
+import { checkDefined } from '../../preconditions/preconditions'
+import { ErrorCode } from './ErrorCode'
 
 const INTERNAL_ERROR = (id?: string) => {
   return {
@@ -23,12 +24,12 @@ const INTERNAL_ERROR = (id?: string) => {
 
 export type APIGatewayProxyHandler = (event: APIGatewayProxyEvent, context: Context) => Promise<APIGatewayProxyResult>
 
-export type ApiRInj = BaseRInj & { requestId: string }
+export type ApiRInj = {
+  log: Logger
+  requestId: string
+}
 
-export type APIHandleRequestParams<CInj, RInj, ReqBody, ReqQueryParams> = BaseHandleRequestParams<
-  CInj,
-  APIGatewayProxyEvent
-> & {
+export type APIHandleRequestParams<CInj, RInj, ReqBody, ReqQueryParams> = {
   context: Context
   event: APIGatewayProxyEvent
   requestBody: ReqBody
@@ -50,9 +51,22 @@ export type ErrorResponse = {
   detail?: string
 }
 
-export abstract class ApiInjector<CInj, RInj extends ApiRInj, ReqBody, ReqQueryParams> extends BaseInjector<CInj> {
-  public constructor(protected injectorName: string) {
-    super(injectorName)
+export abstract class ApiInjector<CInj, RInj extends ApiRInj, ReqBody, ReqQueryParams> {
+  protected containerInjected: CInj | undefined
+
+  public constructor(protected readonly injectorName: string) {
+    checkDefined(injectorName, 'Injector name must be defined')
+  }
+
+  protected abstract buildContainerInjected(): Promise<CInj>
+
+  public async build() {
+    this.containerInjected = await this.buildContainerInjected()
+    return this
+  }
+
+  public getContainerInjected(): CInj {
+    return checkDefined(this.containerInjected, 'Container injected undefined. Must call build() before using.')
   }
 
   public abstract getRequestInjected(
@@ -66,23 +80,11 @@ export abstract class ApiInjector<CInj, RInj extends ApiRInj, ReqBody, ReqQueryP
   ): Promise<RInj>
 }
 
-export abstract class APIGLambdaHandler<
-  CInj,
-  RInj extends ApiRInj,
-  ReqBody,
-  ReqQueryParams,
-  Res
-> extends BaseLambdaHandler<
-  APIGatewayProxyHandler,
-  APIHandleRequestParams<CInj, RInj, ReqBody, ReqQueryParams>,
-  Response<Res> | ErrorResponse
-> {
+export abstract class APIGLambdaHandler<CInj, RInj extends ApiRInj, ReqBody, ReqQueryParams, Res> {
   constructor(
-    handlerName: string,
+    protected readonly handlerName: string,
     private readonly injectorPromise: Promise<ApiInjector<CInj, RInj, ReqBody, ReqQueryParams>>
-  ) {
-    super(handlerName)
-  }
+  ) {}
 
   get handler(): APIGatewayProxyHandler {
     return async (event: APIGatewayProxyEvent, context: Context): Promise<APIGatewayProxyResult> => {
