+++ lib/handlers/get-limit-orders/index.ts
@@ -1,0 +1,9 @@
+import { GetOrdersHandler } from '../get-orders/handler'
+import { GetLimitOrdersInjector } from './injector'
+
+const getLimitOrdersInjectorPromise = new GetLimitOrdersInjector('getLimitOrdersInjector').build()
+const getLimitOrdersHandler = new GetOrdersHandler('getLimitOrdersHandler', getLimitOrdersInjectorPromise)
+
+module.exports = {
+  getLimitOrdersHandler: getLimitOrdersHandler.handler,
+}
