+++ lib/handlers/get-nonce/index.ts
@@ -1,0 +1,9 @@
+import { GetNonceHandler } from './handler'
+import { GetNonceInjector } from './injector'
+
+const getNonceInjectorPromise = new GetNonceInjector('getNonceInjector').build()
+const getNonceHandler = new GetNonceHandler('getNonceHandler', getNonceInjectorPromise)
+
+module.exports = {
+  getNonceHandler: getNonceHandler.handler,
+}
