+++ lib/handlers/get-orders/index.ts
@@ -1,0 +1,8 @@
+import { GetOrdersHandler } from './handler'
+import { GetOrdersInjector } from './injector'
+
+const getOrdersInjectorPromise = new GetOrdersInjector('getOrdersInjector').build()
+const getOrdersHandler = new GetOrdersHandler('getOrdersHandler', getOrdersInjectorPromise)
+module.exports = {
+  getOrdersHandler: getOrdersHandler.handler,
+}
