+++ lib/handlers/order-notification/index.ts
@@ -1,0 +1,12 @@
+import { OrderNotificationHandler } from './handler'
+import { OrderNotificationInjector } from './injector'
+
+const orderNotificationInjectorPromise = new OrderNotificationInjector('orderNotificationInjector').build()
+const orderNotificationHandler = new OrderNotificationHandler(
+  'orderNotificationHandler',
+  orderNotificationInjectorPromise
+)
+
+module.exports = {
+  orderNotificationHandler: orderNotificationHandler.handler,
+}
