+++ lib/handlers/post-order/index.ts
@@ -1,0 +1,9 @@
+import { PostOrderHandler } from './handler'
+import { PostOrderInjector } from './injector'
+
+const postOrderInjectorPromise = new PostOrderInjector('postOrderInjector').build()
+const postOrderHandler = new PostOrderHandler('postOrdersHandler', postOrderInjectorPromise)
+
+module.exports = {
+  postOrderHandler: postOrderHandler.handler,
+}
